requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
google-auth==2.37.0
google-api-python-client==2.159.0
python-multipart==0.0.20
numpy==2.2.1
PyJWT==2.10.1
aiofiles==24.1.0
httpx==0.28.1
keras
tensorflow
